#1 [refactor] application 모듈 삭제
