+++ enterpark-ticket-apis/enduser/src/main/kotlin/com/example/enterparkticket/apis/enduser/controller/user/UserController.kt
@@ -3,7 +3,7 @@ package com.example.enterparkticket.apis.enduser.controller.user
 import com.example.enterparkticket.apis.enduser.controller.user.dto.request.UpdateUserRequest
 import com.example.enterparkticket.apis.enduser.security.dto.AuthUser
 import com.example.enterparkticket.apis.enduser.security.dto.UserPrincipal
-import com.example.enterparkticket.application.usecase.user.command.UserCommandHandler
+import com.example.enterparkticket.domain.user.command.UserCommandHandler
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
@@ -22,7 +22,7 @@ class UserController(private val userCommandHandler: UserCommandHandler) {
         @AuthUser user: UserPrincipal,
         @Valid @RequestBody request: UpdateUserRequest,
     ): ResponseEntity<String> {
-        userCommandHandler.updateUserAddress(user.userId, request.toUpdateUserDto())
+        userCommandHandler.updateUserAddress(user.userId, request.toUpdateUserAddressDto())
         return ResponseEntity.ok("회원 정보 수정이 완료되었습니다.")
     }
 }
