+++ enterpark-ticket-apis/enduser/src/main/kotlin/com/example/enterparkticket/apis/enduser/feign/adapter/KakaoFeignAdapter.kt
@@ -2,8 +2,8 @@ package com.example.enterparkticket.apis.enduser.feign.adapter
 
 import com.example.enterparkticket.apis.enduser.feign.client.KakaoUserClient
 import com.example.enterparkticket.apis.enduser.feign.properties.KakaoProperties
-import com.example.enterparkticket.application.usecase.user.dto.OAuth2UserInfoDto
-import com.example.enterparkticket.application.usecase.user.port.OAuth2Port
+import com.example.enterparkticket.domain.user.command.dto.OAuth2UserInfoDto
+import com.example.enterparkticket.domain.user.port.OAuth2Port
 import com.example.enterparkticket.domain.common.consts.EnterparkTicketConsts.BEARER
 import com.example.enterparkticket.domain.common.consts.EnterparkTicketConsts.KAKAO_AK
 import org.springframework.stereotype.Component
