+++ enterpark-ticket-apis/enduser/src/main/kotlin/com/example/enterparkticket/apis/enduser/feign/dto/response/KakaoTokenResponse.kt
@@ -1,6 +1,5 @@
 package com.example.enterparkticket.apis.enduser.feign.dto.response
 
-import com.example.enterparkticket.application.usecase.user.dto.OAuth2AccessTokenDto
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 
@@ -11,9 +10,4 @@ data class KakaoTokenResponse(
     val expiresIn: Int,
     val refreshToken: String,
     val refreshTokenExpiresIn: Int,
-) {
-
-    fun toOAuthTokenDto(): OAuth2AccessTokenDto {
-        return OAuth2AccessTokenDto(accessToken, expiresIn)
-    }
-}
+)
