--- enterpark-ticket-application/build.gradle.kts
@@ -1,11 +1,0 @@
-import org.springframework.boot.gradle.tasks.bundling.BootJar
-
-val jar: Jar by tasks
-val bootJar: BootJar by tasks
-
-bootJar.enabled = false
-jar.enabled = true
-
-dependencies {
-    implementation(project(":enterpark-ticket-domain"))
-}
