--- enterpark-ticket-application/src/main/kotlin/com/example/enterparkticket/application/usecase/user/dto/OAuth2AccessTokenDto.kt
@@ -1,6 +1,0 @@
-package com.example.enterparkticket.application.usecase.user.dto
-
-data class OAuth2AccessTokenDto(
-    val accessToken: String,
-    val expiresIn: Int,
-)
