--- enterpark-ticket-application/src/main/kotlin/com/example/enterparkticket/application/usecase/user/dto/UpdateUserDto.kt
@@ -1,5 +1,0 @@
-package com.example.enterparkticket.application.usecase.user.dto
-
-data class UpdateUserDto(
-    val address: String,
-)
