--- enterpark-ticket-application/src/main/kotlin/com/example/enterparkticket/application/usecase/user/port/OAuth2Port.kt
@@ -1,10 +1,0 @@
-package com.example.enterparkticket.application.usecase.user.port
-
-import com.example.enterparkticket.application.usecase.user.dto.OAuth2UserInfoDto
-
-interface OAuth2Port {
-
-    fun getOAuth2UserInfo(accessToken: String): OAuth2UserInfoDto
-
-    fun unlinkUser(oAuthId: Long)
-}
