+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/user/command/dto/UpdateUserAddressDto.kt
@@ -1,0 +1,5 @@
+package com.example.enterparkticket.domain.user.command.dto
+
+data class UpdateUserAddressDto(
+    val address: String,
+)
