+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/user/command/dto/UpdateUserDto.kt
@@ -1,4 +1,4 @@
-package com.example.enterparkticket.domain.user.command
+package com.example.enterparkticket.domain.user.command.dto
 
 data class UpdateUserDto(
     val name: String,
