+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/user/port/OAuth2Port.kt
@@ -1,0 +1,10 @@
+package com.example.enterparkticket.domain.user.port
+
+import com.example.enterparkticket.domain.user.command.dto.OAuth2UserInfoDto
+
+interface OAuth2Port {
+
+    fun getOAuth2UserInfo(accessToken: String): OAuth2UserInfoDto
+
+    fun unlinkUser(oAuthId: Long)
+}
