--- enterpark-ticket-infra/build.gradle.kts
@@ -1,15 +1,0 @@
-import org.springframework.boot.gradle.tasks.bundling.BootJar
-
-val jar: Jar by tasks
-val bootJar: BootJar by tasks
-
-bootJar.enabled = false
-jar.enabled = true
-
-dependencies {
-    implementation("org.springframework.boot:spring-boot-starter-data-redis")
-    implementation("org.redisson:redisson-spring-boot-starter:3.40.0")
-
-    implementation(project(":enterpark-ticket-domain"))
-    implementation(project(":enterpark-ticket-application"))
-}
