+++ settings.gradle.kts
@@ -6,10 +6,8 @@ rootProject.name = "enterparkticket"
 include(
     "enterpark-ticket-admin",
     "enterpark-ticket-apis",
-    "enterpark-ticket-application",
     "enterpark-ticket-apis:creator",
     "enterpark-ticket-apis:enduser",
     "enterpark-ticket-batch",
     "enterpark-ticket-domain",
-    "enterpark-ticket-infra",
 )
